+++ packages/terra-search-field/CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 Unreleased
 ----------
 ### Changed
+* Updated search-field example code to show more usage.
 * Components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
 * Update jest test snanpshot
 * updated package.json test scripts